Se agrega nueva estructura refeactorizada
