+++ lib/features/investment/data/models/investment.freezed.dart
@@ -1,0 +1,232 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'investment.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+Investment _$InvestmentFromJson(Map<String, dynamic> json) {
+  return _Investment.fromJson(json);
+}
+
+/// @nodoc
+mixin _$Investment {
+  String get cardId => throw _privateConstructorUsedError;
+  String get cardName => throw _privateConstructorUsedError;
+  String get cardType => throw _privateConstructorUsedError;
+  double get investmentAmount => throw _privateConstructorUsedError;
+
+  /// Serializes this Investment to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of Investment
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $InvestmentCopyWith<Investment> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $InvestmentCopyWith<$Res> {
+  factory $InvestmentCopyWith(
+          Investment value, $Res Function(Investment) then) =
+      _$InvestmentCopyWithImpl<$Res, Investment>;
+  @useResult
+  $Res call(
+      {String cardId,
+      String cardName,
+      String cardType,
+      double investmentAmount});
+}
+
+/// @nodoc
+class _$InvestmentCopyWithImpl<$Res, $Val extends Investment>
+    implements $InvestmentCopyWith<$Res> {
+  _$InvestmentCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of Investment
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? cardId = null,
+    Object? cardName = null,
+    Object? cardType = null,
+    Object? investmentAmount = null,
+  }) {
+    return _then(_value.copyWith(
+      cardId: null == cardId
+          ? _value.cardId
+          : cardId // ignore: cast_nullable_to_non_nullable
+              as String,
+      cardName: null == cardName
+          ? _value.cardName
+          : cardName // ignore: cast_nullable_to_non_nullable
+              as String,
+      cardType: null == cardType
+          ? _value.cardType
+          : cardType // ignore: cast_nullable_to_non_nullable
+              as String,
+      investmentAmount: null == investmentAmount
+          ? _value.investmentAmount
+          : investmentAmount // ignore: cast_nullable_to_non_nullable
+              as double,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$InvestmentImplCopyWith<$Res>
+    implements $InvestmentCopyWith<$Res> {
+  factory _$$InvestmentImplCopyWith(
+          _$InvestmentImpl value, $Res Function(_$InvestmentImpl) then) =
+      __$$InvestmentImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {String cardId,
+      String cardName,
+      String cardType,
+      double investmentAmount});
+}
+
+/// @nodoc
+class __$$InvestmentImplCopyWithImpl<$Res>
+    extends _$InvestmentCopyWithImpl<$Res, _$InvestmentImpl>
+    implements _$$InvestmentImplCopyWith<$Res> {
+  __$$InvestmentImplCopyWithImpl(
+      _$InvestmentImpl _value, $Res Function(_$InvestmentImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of Investment
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? cardId = null,
+    Object? cardName = null,
+    Object? cardType = null,
+    Object? investmentAmount = null,
+  }) {
+    return _then(_$InvestmentImpl(
+      cardId: null == cardId
+          ? _value.cardId
+          : cardId // ignore: cast_nullable_to_non_nullable
+              as String,
+      cardName: null == cardName
+          ? _value.cardName
+          : cardName // ignore: cast_nullable_to_non_nullable
+              as String,
+      cardType: null == cardType
+          ? _value.cardType
+          : cardType // ignore: cast_nullable_to_non_nullable
+              as String,
+      investmentAmount: null == investmentAmount
+          ? _value.investmentAmount
+          : investmentAmount // ignore: cast_nullable_to_non_nullable
+              as double,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$InvestmentImpl implements _Investment {
+  _$InvestmentImpl(
+      {required this.cardId,
+      required this.cardName,
+      required this.cardType,
+      required this.investmentAmount});
+
+  factory _$InvestmentImpl.fromJson(Map<String, dynamic> json) =>
+      _$$InvestmentImplFromJson(json);
+
+  @override
+  final String cardId;
+  @override
+  final String cardName;
+  @override
+  final String cardType;
+  @override
+  final double investmentAmount;
+
+  @override
+  String toString() {
+    return 'Investment(cardId: $cardId, cardName: $cardName, cardType: $cardType, investmentAmount: $investmentAmount)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$InvestmentImpl &&
+            (identical(other.cardId, cardId) || other.cardId == cardId) &&
+            (identical(other.cardName, cardName) ||
+                other.cardName == cardName) &&
+            (identical(other.cardType, cardType) ||
+                other.cardType == cardType) &&
+            (identical(other.investmentAmount, investmentAmount) ||
+                other.investmentAmount == investmentAmount));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode =>
+      Object.hash(runtimeType, cardId, cardName, cardType, investmentAmount);
+
+  /// Create a copy of Investment
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$InvestmentImplCopyWith<_$InvestmentImpl> get copyWith =>
+      __$$InvestmentImplCopyWithImpl<_$InvestmentImpl>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$InvestmentImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _Investment implements Investment {
+  factory _Investment(
+      {required final String cardId,
+      required final String cardName,
+      required final String cardType,
+      required final double investmentAmount}) = _$InvestmentImpl;
+
+  factory _Investment.fromJson(Map<String, dynamic> json) =
+      _$InvestmentImpl.fromJson;
+
+  @override
+  String get cardId;
+  @override
+  String get cardName;
+  @override
+  String get cardType;
+  @override
+  double get investmentAmount;
+
+  /// Create a copy of Investment
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$InvestmentImplCopyWith<_$InvestmentImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
