+++ lib/features/investment/domain/datasource/investment_datasource.dart
@@ -1,0 +1,6 @@
+import '../../data/models/investment.dart';
+
+abstract class InvestmentDataSource {
+  Future<List<Investment>> getAllInvestments();
+  Future<void> addInvestment(Investment investment);
+}
