+++ lib/features/investment/domain/repository/investment_repository.dart
@@ -1,0 +1,6 @@
+import '../../data/models/investment.dart';
+
+abstract class InvestmentRepository {
+  Future<List<Investment>> getAllInvestments();
+  Future<void> addInvestment(Investment investment);
+}
