+++ pubspec.yaml
@@ -29,14 +29,12 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   freezed_annotation: ^2.4.1
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   dio: ^5.2.0
   flutter_riverpod: ^2.6.1
   riverpod_annotation: ^2.0.0
   flutter:
     sdk: flutter
-    assets:
-    - lib/assets/investment.json
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -50,8 +48,6 @@ dev_dependencies:
   json_serializable: ^6.7.1
   flutter_test:
     sdk: flutter
-    assets:
-    - lib/assets/investment.json
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
